requirements.txt
# FastAPI core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.0

# HTTP requests
httpx>=0.25.0
requests>=2.31.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
yfinance>=0.2.28

# Environment
python-dotenv>=1.0.0

# Plotting (optional)
plotly>=5.17.0

# Text processing (optional)
textblob>=0.17.1
vaderSentiment>=3.3.2

# Utilities
python-multipart>=0.0.6
email-validator>=2.1.0

# AI Trading Agent - Alpaca Integration
alpaca-py>=0.21.0
websockets>=12.0
httpx>=0.25.0

# AI Trading Agent - Claude API
anthropic>=0.18.0

# AI Trading Agent - CLI Dashboard
rich>=13.7.0

# AI Trading Agent - Async support
aiohttp>=3.9.0
asyncio-throttle>=1.0.0

# Supabase (log storage + frontend auth)
supabase>=2.3.0